src/backend/tests/requirements.txt
# Test requirements for agentic retrieval tests
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0
requests>=2.31.0
azure-identity>=1.14.0
semantic-kernel>=0.9.0